Start of temp read

+++ include/IOLMasterPort.h
@@ -79,7 +79,7 @@ public:
 
     virtual void writeISDU() = 0;
 
-	virtual uint8_t readDirectParameterPage(uint8_t address, uint8_t *pData) = 0;
+	virtual uint8_t readDirectParameterPage(uint16_t address, uint8_t *pData) = 0;
 
     virtual uint8_t readPD(uint8_t *pData, uint8_t sizeData) = 0;
