+++ include/IOLMasterPortMax14819.h
@@ -82,7 +82,7 @@ public:
 
 	void writeISDU();
 
-	uint8_t readDirectParameterPage(uint8_t address, uint8_t *pData);
+	uint8_t readDirectParameterPage(uint16_t address, uint8_t *pData);
 
 	uint8_t readPD(uint8_t *pData, uint8_t sizeData);
